sign_dataset/backend/requirements.txt
Flask-Cors==3.0.8
Flask==1.1.2
Werkzeug==0.16.1
tensorflow==1.14.0
numpy
Pillow
gunicorn==20.0.4
